refactor(): further simplification & cleanups
